Controlled Serial for debug, Some Refactoring
I moved all interfaces to an interfaces folder under infrastructure. I added the IControlledSerial which is an interface to the Serial object for Arduino, allowing Message queueing, and asynchronous printing. (That is, messages needn't be printed at the creating time.) This will allow us to separate our debugging messages from the timing of the program without other problems.

I corrected various other files to accomodate the refactoring changes, and I added the controlled Serial to the main.cpp for use. Just call the print messages at the end of Loop().

WARNING: At present, I do not know if there will be any issues between the ISR blinker and the Serial Print messages. I may need to add some interupt safety to it. However that will require more research.